pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradingview-schwab-bridge"
version = "0.1.0"
description = "Webhook bridge from TradingView alerts to Schwab market orders"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.7,<3.0",
    "schwab-py>=1.5.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
